Move Events/EventFilter to their own files
