+++ src/parser/core/Events.js
@@ -1,0 +1,12 @@
+import EventFilter from './EventFilter';
+
+const Events = {
+  get cast() {
+    return new EventFilter('cast');
+  },
+  get heal() {
+    return new EventFilter('heal');
+  },
+};
+
+export default Events;
